move logic of coffee to CoffeeService

+++ src/main/java/band/effective/coffeeshop/model/Ingredient.java
@@ -7,6 +7,7 @@ import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotEmpty;
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.util.Set;
 //many-to-many problem
 @Data
@@ -34,7 +35,7 @@ public class Ingredient {
     @NonNull
     @Min(0)
     @JsonAlias({"price_per_one", "cost_per_one"})
-    private Double costPerOne;
+    private BigDecimal costPerOne;
 
 
 }
